Refactored Payloads to allow for adding context to receive (e.g. downloading by client)

+++ storage-core/src/common/messages/chunk_transfer.rs
@@ -10,15 +10,29 @@ use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt, BufWriter};
 pub struct ChunkTransfer {
     pub offset: Option<u64>,
     pub data: PathBuf,
+    should_delete: bool,
 }
 
 impl ChunkTransfer {
+    pub fn new(offset: Option<u64>, data: PathBuf) -> Self {
+        ChunkTransfer {
+            offset,
+            data,
+            should_delete: false,
+        }
+    }
+
+    pub fn commit(mut self) -> PathBuf {
+        self.should_delete = false;
+        self.data.clone()
+    }
+
     pub(crate) async fn send_chunk(
         &self,
         chunk_size: u64,
         send: &mut SendStream,
     ) -> anyhow::Result<()> {
-        let mut file = tokio::fs::File::open(&self.data).await?;
+        let mut file = File::open(&self.data).await?;
 
         if let Some(offset) = self.offset {
             AsyncSeekExt::seek(&mut file, SeekFrom::Start(offset)).await?;
@@ -36,31 +50,22 @@ impl ChunkTransfer {
     }
 
     pub(crate) async fn recv_chunk(
-        chunk_id: ChunkId,
+        &self,
         chunk_size: u64,
+        file_writer: &mut BufWriter<File>,
         recv: &mut RecvStream,
-    ) -> anyhow::Result<Self> {
-        let data = TMP_STORAGE_ROOT
-            .get()
-            .expect("Temporary storage not initialized via config")
-            .join(chunk_id.to_string());
-
-        let file = File::create(&data).await?;
-        file.set_len(chunk_size).await?;
-        let mut writer = BufWriter::with_capacity(chunk_size as usize, file);
+    ) -> anyhow::Result<()> {
         let mut limited_recv = recv.take(chunk_size);
-        tokio::io::copy(&mut limited_recv, &mut writer).await?;
-        writer.flush().await?;
+        tokio::io::copy(&mut limited_recv, file_writer).await?;
+        file_writer.flush().await?;
 
-        writer.into_inner().sync_all().await?;
-
-        Ok(ChunkTransfer { data, offset: None })
+        Ok(())
     }
 }
 
 impl Drop for ChunkTransfer {
     fn drop(&mut self) {
-        if self.data.exists() {
+        if self.should_delete && self.data.exists() {
             let _ = std::fs::remove_file(&self.data);
         }
     }
